refactor: Give multi-word names to some components
