+++ components/timer/splashScreen.vue
